+++ cmd/in.go
@@ -16,21 +16,16 @@ package cmd
 
 import (
 	"fmt"
-	"io"
-	"os"
 	"time"
 
 	"github.com/spf13/viper"
 
 	"github.com/lucassabreu/clockify-cli/api/dto"
-	"github.com/lucassabreu/clockify-cli/reports"
 
 	"github.com/lucassabreu/clockify-cli/api"
 	"github.com/spf13/cobra"
 )
 
-var cardNumber int
-var issueNumber int
 var tags []string
 var notBillable bool
 var noClosing bool
@@ -41,10 +36,10 @@ var whenToCloseString string
 
 // inCmd represents the in command
 var inCmd = &cobra.Command{
-	Use:     "in <project-name-or-id> <description>",
-	Short:   "Create a new time entry and starts it",
-	Example: `clockify-cli in --issue 13 "time sheet"`,
+	Use:     "in <project-id> <description>",
+	Short:   "Create a new time entry and starts it (will close time entries not closed)",
 	Args:    cobra.MaximumNArgs(2),
+	Aliases: []string{"start"},
 	Run: withClockifyClient(func(cmd *cobra.Command, args []string, c *api.Client) {
 
 		var whenToCloseDate time.Time
@@ -81,35 +76,10 @@ var inCmd = &cobra.Command{
 			tei.TimeInterval.End = &whenToCloseDate
 		}
 
-		tei, err = newEntry(c, tei, viper.GetBool("interactive"), !noClosing)
-
-		if err != nil {
-			printError(err)
-			return
-		}
-
-		te, err := c.ConvertIntoFullTimeEntry(tei)
-		if err != nil {
-			printError(err)
-			return
-		}
-
 		format, _ := cmd.Flags().GetString("format")
 		asJSON, _ := cmd.Flags().GetBool("json")
-
-		var reportFn func(*dto.TimeEntry, io.Writer) error
-
-		reportFn = reports.TimeEntryPrint
-
-		if asJSON {
-			reportFn = reports.TimeEntryJSONPrint
-		}
-
-		if format != "" {
-			reportFn = reports.TimeEntryPrintWithTemplate(format)
-		}
-
-		if err = reportFn(&te, os.Stdout); err != nil {
+		err = newEntry(c, tei, viper.GetBool("interactive"), true, format, asJSON)
+		if err != nil {
 			printError(err)
 		}
 	}),
@@ -120,18 +90,14 @@ func init() {
 
 	addTimeEntryFlags(inCmd)
 
-	inCmd.Flags().StringP("format", "f", "", "golang text/template format to be applyed on each time entry")
+	inCmd.Flags().StringP("format", "f", "", "golang text/template format to be applied on each time entry")
 	inCmd.Flags().BoolP("json", "j", false, "print as json")
-	inCmd.PersistentFlags().BoolVar(&noClosing, "no-closing", false, "don't close any active time entry")
-
-	_ = viper.BindPFlag("no-closing", inCmd.PersistentFlags().Lookup("no-closing"))
 }
 
 func addTimeEntryFlags(cmd *cobra.Command) {
 	cmd.Flags().BoolVarP(&notBillable, "not-billable", "n", false, "is this time entry not billable")
-	cmd.Flags().IntVarP(&cardNumber, "card", "c", 0, "trello card number being started")
-	cmd.Flags().IntVar(&issueNumber, "issue", 0, "issue number being started")
 	cmd.Flags().StringVar(&task, "task", "", "add a task to the entry")
 	cmd.Flags().StringSliceVar(&tags, "tag", []string{}, "add tags to the entry")
-	cmd.Flags().StringVar(&whenString, "when", time.Now().Format(fullTimeFormat), "when the entry should be closed, if not informed will use current time")
+	cmd.Flags().StringVar(&whenString, "when", time.Now().Format(fullTimeFormat), "when the entry should be started, if not informed will use current time")
+	cmd.Flags().StringVar(&whenToCloseString, "when-to-close", "", "when the entry should be closed, if not informed will let it open")
 }