+++ cmd/manual.go
@@ -1,0 +1,84 @@
+// Copyright © 2020 Lucas dos Santos Abreu <lucas.s.abreu@gmail.com>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/lucassabreu/clockify-cli/api"
+	"github.com/lucassabreu/clockify-cli/api/dto"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+// manualCmd represents the manual command
+var manualCmd = &cobra.Command{
+	Use:   "manual <project-id> <start> <end> <description>",
+	Short: "Creates a new completed time entry (does not stop on-going time entries)",
+	Args:  cobra.MaximumNArgs(4),
+	Run: withClockifyClient(func(cmd *cobra.Command, args []string, c *api.Client) {
+		var whenToCloseDate time.Time
+		var err error
+
+		tei := dto.TimeEntryImpl{
+			WorkspaceID:  viper.GetString("workspace"),
+			TagIDs:       tags,
+			TimeInterval: dto.TimeInterval{},
+		}
+
+		if len(args) > 0 {
+			tei.ProjectID = args[0]
+		}
+
+		if len(args) > 1 {
+			tei.TimeInterval.Start, err = convertToTime(args[1])
+			if err != nil {
+				printError(fmt.Errorf("Fail to convert when to start: %s", err.Error()))
+				return
+			}
+		}
+
+		if len(args) > 2 {
+			whenToCloseDate, err = convertToTime(args[2])
+			if err != nil {
+				printError(fmt.Errorf("Fail to convert when to end: %s", err.Error()))
+				return
+			}
+			tei.TimeInterval.End = &whenToCloseDate
+		}
+
+		if len(args) > 3 {
+			tei.Description = args[3]
+		}
+
+		format, _ := cmd.Flags().GetString("format")
+		asJSON, _ := cmd.Flags().GetBool("json")
+		err = newEntry(c, tei, viper.GetBool("interactive"), false, format, asJSON)
+		if err != nil {
+			printError(err)
+		}
+	}),
+}
+
+func init() {
+	rootCmd.AddCommand(manualCmd)
+
+	manualCmd.Flags().BoolVarP(&notBillable, "not-billable", "n", false, "is this time entry not billable")
+	manualCmd.Flags().StringVar(&task, "task", "", "add a task to the entry")
+	manualCmd.Flags().StringSliceVar(&tags, "tag", []string{}, "add tags to the entry")
+	manualCmd.Flags().StringP("format", "f", "", "golang text/template format to be applied on each time entry")
+	manualCmd.Flags().BoolP("json", "j", false, "print as json")
+}